backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0

# AI / Gemini
google-genai>=0.3.0

# File Processing (PDF, DOCX, Markdown)
pypdf>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdown>=3.5.0

# Knowledge Base / RAG
numpy>=1.24.0

# Web Scraping
requests>=2.31.0
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
pydantic>=2.0.0
python-multipart>=0.0.6
